requirements.txt
flask==3.0.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
prometheus-client==0.19.0
requests==2.31.0
joblib==1.3.2
schedule==1.2.0
pytest==7.4.3
python-dotenv==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
psutil==5.9.6
pyyaml==6.0.1
